requirements.txt
pyspark
apache-airflow
neo4j
pandas
streamlit
pytest
requests
pyyaml